pyproject.toml
[project]
name = "delegate-ai"
version = "0.1.0"
description = "Agentic team management system"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.40.0",
    "claude-code-sdk>=0.0.25",
    "click>=8.1",
    "fastapi>=0.128.4",
    "pyyaml>=6.0.3",
    "uvicorn[standard]>=0.40.0",
    "watchfiles>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["delegate"]

# Run frontend build (esbuild) before creating the wheel
[tool.hatch.build.hooks.custom]

[project.scripts]
delegate = "delegate.cli:main"

[dependency-groups]
dev = ["httpx>=0.28.1", "pytest>=9.0.2", "pytest-asyncio>=1.3.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
